refactor: ShortestPathEdge 패키지 변경
