+++ README.md
@@ -35,7 +35,7 @@
 - [x] 접근 제한자 확인 및 final 선언
 - [x] Fake 객체를 이용 -> test DB로 이전
 - [x] fare + test 가독성 높이는 방향으로 수정
-- [ ] exception -> state code 별로 구분
+- [x] exception -> state code 별로 구분
 - [ ] LineService Test 상수 전체 적용 및 검증부 수정
 - [x] FareBy test 예외 케이스 추가
 - [ ] Path 에서 fare 분리
